Refactored storage.py to be compatible with snowflake instead of SQLite.
As of right now, this does not work because Snowflake student accounts
require Multifactor Authentication (MFA) which cannot be done through the
snowflake-connector-python library. Further work is needed to find a workaround.

It may be possible to get Gil to disable MFA for student accounts, or to
create a standard account. I have not reached out to Gil, but I did attempt
to create a standard account myself, but was unsuccessful so far. Snowflake
won't send me my confirmation email, and has no form to request a new one.